pyproject.toml
[project]
name = "the-amazing-agent-race"
version = "0.1.0"
description = "The Amazing Agent Race — a benchmark for evaluating LLM agents on multi-step scavenger-hunt puzzles"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai",
    "python-dotenv",
    "httpx",
    "markdownify",
    "mcp",
    "anyio",
    "click",
    "yfinance",
    "matplotlib>=3.10.8",
    "datasets>=4.8.4",
    "huggingface-hub>=1.8.0",
]
